Revert "Chempred pub"
