+++ README.md
@@ -1,8 +1,2 @@
 # sciNER
-Character-level RNN for named entity recognition in plain text.
-
-This is a special ChemPred publication branch of the sciNER package. ChemPred
-is a deep CNN-RNN architecture for chemical NER trained on the CHEMDNER corpus
-(publication pending). Jupyter notebook `chempred-example.ipynb` will guide
-you through the steps needed to train and validate the model. You can also find
-trained weights for the model in `./models/chempred-model-weights.hdf5`.
+Character-level RNN for named entity recognition in plain text
