--- sciner/config.py
@@ -1,53 +1,0 @@
-import json
-from io import TextIOWrapper
-from typing import Union, Mapping
-
-
-class Config(dict):
-    """
-    Model configurations
-    """
-    def __init__(self, config: Union[TextIOWrapper, Mapping]):
-        """
-        :param config: an opened json file or a mapping
-        """
-        super(Config, self).__init__(config if isinstance(config, Mapping) else
-                                     json.load(config))
-
-    def __getitem__(self, item):
-        retval = self.get(item)
-        if retval is None:
-            raise KeyError("No {} configuration".format(item))
-        return retval
-
-    def get(self, item, default=None):
-        """
-        :param item: item to search for
-        :param default: default value
-        :return:
-        >>> config = Config(open("testdata/config-detector.json"))
-        >>> config["bidirectional"]
-        True
-        >>> config.get("epochs")
-        >>> from_dict = Config(json.load(open("testdata/config-detector.json")))
-        >>> from_mapping = Config(from_dict)
-        >>> from_mapping["nsteps"]
-        [200, 200, 200, 200]
-        >>> from_mapping.update({"lstm": {"nsteps": [300, 300]}})
-        >>> isinstance(from_mapping, Config)
-        True
-        >>> from_mapping["nsteps"]
-        [300, 300]
-        """
-        to_visit = list(self.items())
-        while to_visit:
-            key, value = to_visit.pop()
-            if key == item:
-                return value
-            if isinstance(value, Mapping):
-                to_visit.extend(value.items())
-        return default
-
-
-if __name__ == "__main__":
-    raise RuntimeError
