+++ sciner/training.py
@@ -1,0 +1,122 @@
+import glob
+import operator as op
+import os
+import re
+import shutil
+from contextlib import contextmanager
+from itertools import starmap, groupby
+from typing import Tuple, List, Iterable, Text, Sequence, \
+    Optional, Mapping, cast
+
+import numpy as np
+from fn import F
+from sklearn.utils import class_weight
+
+from sciner.intervals import Interval, Intervals, extract
+from sciner.sampling import Annotator, Sampler, Sample
+
+ProcessedSample = Tuple[int, Text, Sequence[Interval], Sequence[Text],
+                        Optional[np.ndarray]]
+
+def group(ids, sources, *args):
+    """
+    Group args by id and source
+    :param ids:
+    :param sources:
+    :param args:
+    :return:
+    """
+    records = zip(ids, sources, *args)
+    id_groups = groupby(records, op.itemgetter(0))
+    return [[list(grp) for _, grp in src_grps] for src_grps in
+            (groupby(list(grp), op.itemgetter(1)) for _, grp in id_groups)]
+
+
+def pick_best(filenames: List[str]) -> Tuple[str, Tuple[int, float]]:
+    # TODO import docs
+    """
+    >>> fnames = ["rootdir/name/weights-improvement-16-0.99.hdf5",
+    ...           "rootdir/name/weights-improvement-25-0.99.hdf5",
+    ...           "rootdir/name/weights-improvement-20-0.99.hdf5"]
+    >>> pick_best(fnames)
+    ('rootdir/name/weights-improvement-25-0.99.hdf5', (25, 0.99))
+    """
+    pattern = re.compile("([0-9]+)-([0-9.]+)\.hdf5")
+    stats = (F(map, pattern.findall)
+             >> (map, op.itemgetter(0))
+             >> (starmap, lambda epoch, acc: (int(epoch), float(acc)))
+             )(filenames)
+    return max(zip(filenames, stats), key=op.itemgetter(1))
+
+
+@contextmanager
+def training(rootdir: str, name: str):
+    # TODO docs
+    """
+    Initialise temporary training directories and cleanup upon completion
+    :param rootdir:
+    :param name:
+    :return:
+    """
+    training_dir = os.path.join(rootdir, "{}-training".format(name))
+    weights_template = os.path.join(training_dir,
+                                    "{epoch:02d}-{val_acc:.3f}.hdf5")
+    destination = os.path.join(rootdir, "{}.hdf5".format(name))
+    os.makedirs(training_dir)
+    try:
+        yield weights_template
+    finally:
+        all_checkpoints = glob.glob(os.path.join(training_dir, "*.hdf5"))
+        if all_checkpoints:
+            best_checkpoint, stats = pick_best(all_checkpoints)
+            shutil.move(best_checkpoint, destination)
+        shutil.rmtree(training_dir)
+
+
+def balance_class_weights(y: np.ndarray, mask: Optional[np.ndarray]=None) \
+        -> Optional[Mapping[int, float]]:
+    # TODO update docs
+    # TODO tests
+    """
+    :param y: a numpy array encoding sample classes; samples are encoded along
+    the 0-axis
+    :param mask: a boolean array of shape compatible with `y`, wherein True
+    shows that the corresponding value(s) in `y` should be used to calculate
+    weights; if `None` the function will consider all values in `y`
+    :return: class weights
+    """
+    if not len(y):
+        raise ValueError("`y` is empty")
+    if y.ndim == 2:
+        y_flat = (y.flatten() if mask is None else
+                  np.concatenate([sample[mask] for sample, mask in zip(y, mask)]))
+    elif y.ndim == 3:
+        y_flat = (y.nonzero()[-1] if mask is None else
+                  y[mask].nonzero()[-1])
+    else:
+        raise ValueError("`y` should be either a 2D or a 3D array")
+    classes = np.unique(y_flat)
+    weights = class_weight.compute_class_weight("balanced", classes, y_flat)
+    weights_scaled = weights / weights.min()
+    return {cls: weight for cls, weight in zip(classes, weights_scaled)}
+
+
+def sample_weights(y: np.ndarray, class_weights: Mapping[int, float]) \
+        -> np.ndarray:
+    # TODO update docs
+    # TODO tests
+    """
+    :param y: a 2D array encoding sample classes; each sample is a row of
+    integers representing class code
+    :param class_weights: a class to weight mapping
+    :return: a 2D array of the same shape as `y`, wherein each position stores
+    a weight for the corresponding position in `y`
+    """
+    weights_mask = np.zeros(shape=y.shape, dtype=np.float32)
+    for cls, weight in class_weights.items():
+        weights_mask[y == cls] = weight
+    return weights_mask
+
+
+if __name__ == "__main__":
+    raise RuntimeError
