+++ test.py
@@ -1,3 +1,4 @@
+import re
 import unittest
 from typing import Sequence, Iterable, cast
 
@@ -5,9 +6,8 @@ import numpy as np
 from hypothesis import given, note
 from hypothesis import settings, strategies as st
 
-from sciner import intervals
-from sciner.corpora import genia
-from sciner.preprocessing import preprocessing
+from sciner import intervals, text, genia, sampling, util
+
 
 MAX_TESTS = 1000
 
@@ -31,12 +31,15 @@ class TestText(unittest.TestCase):
             codes[iv.start:iv.stop] = list(map(ord, token))
         return "".join(map(chr, codes))
 
-    # @given(texts)
-    # @settings(max_examples=MAX_TESTS)
-    # def test_parse_text(self, txt):
-    #     parsed = text.tointervals(text.fine_tokeniser, txt)
-    #     mod_text = re.sub("\s", " ", txt)
-    #     self.assertEqual(self.unparse(txt, parsed), mod_text.rstrip())
+    @given(texts)
+    @settings(max_examples=MAX_TESTS)
+    def test_parse_text(self, txt):
+        parsed = text.tointervals(text.spacy_tokeniser, txt)
+        mod_text = re.sub("\s", " ", txt)
+        self.assertEqual(self.unparse(txt, parsed), mod_text.rstrip())
+        parsed = text.tointervals(text.fine_tokeniser, txt)
+        mod_text = re.sub("\s", " ", txt)
+        self.assertEqual(self.unparse(txt, parsed), mod_text.rstrip())
 
 
 class TestGenia(unittest.TestCase):
@@ -69,27 +72,27 @@ class TestSampling(unittest.TestCase):
         ivs = [intervals.Interval(arr[0], arr[-1]+1) for arr in
                np.split(np.arange(length), split_points)[1:-1]]
 
-        sample_anno = preprocessing.annotate_sample(ncls, anno, ivs)
+        sample_anno = sampling.annotate_sample(ncls, anno, ivs)
         sample_anno_cls = [set(iv_anno.nonzero()[-1])
                            for iv_anno in cast(Iterable[np.ndarray], sample_anno)]
         self.assertSequenceEqual([set(anno[iv.start:iv.stop]) for iv in ivs],
                                  sample_anno_cls)
         self.assertEqual(len(ivs), len(sample_anno))
 
-    # @given(st.integers(1, 10), st.integers(1, 100), st.integers(0, 100))
-    # @settings(max_examples=MAX_TESTS)
-    # def test_flatten_multilabel_annotation(self, ncls, nsteps, maxmixed):
-    #     anno = np.random.choice(ncls, nsteps)
-    #     nested = util.one_hot(ncls, anno)
-    #     mixed_steps = np.random.choice(nsteps, min(nsteps, maxmixed), False)
-    #     nested[mixed_steps, 0] = 1
-    #     self.assertTrue(
-    #         (anno == sampling.flatten_multilabel_annotation(nested)).all()
-    #     )
-    #     nested[mixed_steps, np.random.choice(ncls)] = 1
-    #     if ncls > 2 and (nested[:, 1:].sum(1) > 1).any():
-    #         with self.assertRaises(sampling.AmbiguousAnnotation):
-    #             sampling.flatten_multilabel_annotation(nested)
+    @given(st.integers(1, 10), st.integers(1, 100), st.integers(0, 100))
+    @settings(max_examples=MAX_TESTS)
+    def test_flatten_multilabel_annotation(self, ncls, nsteps, maxmixed):
+        anno = np.random.choice(ncls, nsteps)
+        nested = util.one_hot(ncls, anno)
+        mixed_steps = np.random.choice(nsteps, min(nsteps, maxmixed), False)
+        nested[mixed_steps, 0] = 1
+        self.assertTrue(
+            (anno == sampling.flatten_multilabel_annotation(nested)).all()
+        )
+        nested[mixed_steps, np.random.choice(ncls)] = 1
+        if ncls > 2 and (nested[:, 1:].sum(1) > 1).any():
+            with self.assertRaises(sampling.AmbiguousAnnotation):
+                sampling.flatten_multilabel_annotation(nested)
 
 
 if __name__ == "__main__":
